requirements.txt
wfdb
PyYAML
pyarrow
py-ecg-detectors
torch